+++ stac-cli/src/error.rs
@@ -1,21 +1,14 @@
-use crate::download::Progress;
 use stac::Value;
 use thiserror::Error;
 
 #[derive(Error, Debug)]
 #[non_exhaustive]
 pub enum Error {
-    #[error("unsupported STAC type for downloading assets")]
-    CannotDownload(Value),
-
     #[error(transparent)]
     Io(#[from] std::io::Error),
 
     #[error("invalid STAC")]
     InvalidValue(Value),
-
-    #[error(transparent)]
-    ProgressSend(#[from] tokio::sync::mpsc::error::SendError<Progress>),
 
     #[error(transparent)]
     StacAsync(#[from] stac_async::Error),