+++ stac-cli/src/lib.rs
@@ -1,10 +1,8 @@
 mod args;
-mod download;
 mod error;
 
 pub use {
     args::{Args, Command},
-    download::download,
     error::Error,
 };